refactor: rewrite 'apiSlice', add constants (numbers, api)
