+++ src/common/constants/api.ts
@@ -1,0 +1,10 @@
+// http://localhost:9011 | https://rs-lang-en.herokuapp.com
+export const BASE_URL = 'https://rs-lang-en.herokuapp.com';
+
+export const REDUCER_PATH = 'api';
+
+export const API = {
+  words: '/words',
+  users: '/users',
+  signin: '/signin',
+};
