+++ src/features/api/apiSlice.ts
@@ -1,29 +1,33 @@
 import { Word } from './../../common/types/interfaces';
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
-import { DifficultyLevel } from '../../common/types/enums';
 import { WordsQuery } from '../../common/types/interfaces';
+import { API, BASE_URL, REDUCER_PATH } from '../../common/constants/api';
+import { DifficultyLevel, INITIAL_VALUE } from '../../common/constants/numbers';
 
-const BASE_URL = 'https://rs-lang-en.herokuapp.com'; // http://localhost:9011 | https://rs-lang-en.herokuapp.com
-
-function groupChecker(group: DifficultyLevel) {
-  if (group < 0) {
-    return 0;
-  } else if (group > 5) {
-    return 5;
+const groupChecker = (group: DifficultyLevel) => {
+  if (group < DifficultyLevel.LEVEL_0) {
+    return DifficultyLevel.LEVEL_0;
+  } else if (group > DifficultyLevel.LEVEL_5) {
+    return DifficultyLevel.LEVEL_5;
   }
   return group;
-}
+};
 
 export const apiSlice = createApi({
-  reducerPath: 'api',
+  reducerPath: REDUCER_PATH,
   baseQuery: fetchBaseQuery({ baseUrl: BASE_URL }),
   endpoints: (builder) => ({
     getWords: builder.query<Word[], WordsQuery>({
-      query: ({ group = DifficultyLevel.LEVEL_0, page = 0 }) =>
-        `/words?group=${groupChecker(group)}&page=${page}`,
+      query: ({ group = DifficultyLevel.LEVEL_0, page = INITIAL_VALUE }) => ({
+        url: API.words,
+        params: {
+          group: groupChecker(group),
+          page: page,
+        },
+      }),
     }),
     getWord: builder.query<Word, string>({
-      query: (wordId: string) => `/words/${wordId}`,
+      query: (wordId: string) => `${API.words}/${wordId}`,
     }),
   }),
 });
